Change cOOre compiler to add local as suffix
See: http://rtfm.codeplex.com/wikipage?title=Bindings1&referringTitle=2nd%20Challenge

Recall, that the local definitions were introduced a while back as a means to run under the claim of the parenting object. Without distinguishing local from ordinary function we would run into the situation where we try to re-claim a lock, that we have already taken (causing a sort of deadlock). We chose to allow re-claiming ONLY the case is stems from a direct synchronous call from within the same object (and hence discarding re-claims occurring in sync chains.)

This leads us into the first discussion on how this can be be achieved by the -cOOre/-core compilers.

Looking deeper at the problem, both -cOOre and -core perform specialisation, (-cOOre on basis of object instances, while -core on basis of task and function instances.)

For this example, the -cOOre compiler will generate an instance of root (of the class definition Root). Any tasks and functions will be pre-pended with root_, e.g., the function f(...) in the root instance is assumed the qualified name root_f(...) and the local function root_f_local(...) accordingly. (Notice the subtle change of post pending local instead of prepending it, the reason will soon be revealed...)

To get this working, we need the collaboration of the -core compiler. We have given the definition for f_local(...), but for the correct composition with the caller assuming root_f_local(...) we need to instruct -core to specialise the inclusion of "ExtRoot.core".